requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.3
djangorestframework==3.12.4
django-cors-headers==3.7.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.4.0